Add files via upload
Alarm clock weather stations which has a remote sensor unit that is placed outside in order to measure meteorological characteristics should as temperature and relative humidity. In this way the user is usefully informed as to the manner in which they should attire themselves prior to leaving the home.
The inputs to the system are simulated by pushbutton switches and the built-in transducers. The LCD is used to display menu options and show essential information including current time, alarm time, outside temperature, and relative humidity. 
LEDs are used to simulate outputs such the alarm buzzer.